Move ErrNotImplemented definition to consensus package
