+++ config/config.go
@@ -58,8 +58,8 @@ var (
 	Default = Config{
 		NodeType: FullNodeType,
 		Network: Network{
-			Host:                                "127.0.0.1",
-			Port:                                4689,
+			Host: "127.0.0.1",
+			Port: 4689,
 			MsgLogsCleaningInterval:             2 * time.Second,
 			MsgLogRetention:                     5 * time.Second,
 			HealthCheckInterval:                 time.Second,
@@ -115,13 +115,13 @@ var (
 				AcceptProposeTTL:         time.Second,
 				AcceptProposalEndorseTTL: time.Second,
 				AcceptCommitEndorseTTL:   time.Second,
-				Delay:                    5 * time.Second,
-				NumSubEpochs:             1,
-				EventChanSize:            10000,
-				NumDelegates:             21,
-				EnableDummyBlock:         true,
-				TimeBasedRotation:        false,
-				EnableDKG:                false,
+				Delay:             5 * time.Second,
+				NumSubEpochs:      1,
+				EventChanSize:     10000,
+				NumDelegates:      21,
+				EnableDummyBlock:  true,
+				TimeBasedRotation: false,
+				EnableDKG:         false,
 			},
 			BlockCreationInterval: 10 * time.Second,
 		},
